Fix HDF5 system-wide search
This is a follow-up to cdb278b, which erroneously assumes that the last
directory of $HDF5_LIBRARY_DIRS is "/lib" which is only true on mac. On debian,
for example, it is "/openmpi" for the parallel HDF5 system wide. So now we only
strip off the "/lib" on mac (as previous to cdb278b).