Rename Conductor to Transporter
This is the first step in refactoring the class names in inciter:

Instead of Conductor we will use Transporter, and instead of Performer we will
use Carrier. These names will better express the fact the Transporter Charm++
chare will conduct a multiple Carrier Charm++ chare array elements, which will
work together to solve a system of transport equations.

This is constrated to Striker and HitMan, which will be analogous to Transporter
and Carrier, and will solve (in a single step/strike) a partial different
equation system that is not a transport equation, and therefore does not require
multiple time steps.