Abstract particle velocity to PDEs
This commit abstracts the previously hard-coded indices of accessing the
cell-nodal velocity field in Carrier to the PDE level. The previously hard-coded
version only worked for CompNS (and Euler) PDEs, however, Carrier can integrate
a system of system of PDEs and this should be kept that way. So this is a step
toward keeping Carrier general.

Design rationale: Since there is a system of system of PDEs being integrated by
the Carrier chares, we loop through all PDEs configured by the user and query
the velocities in a mesh cell. Now some PDE systems can provide this velocity
because they have a well defeined velocity field, e.g., Euler, CompNS, but some
others, e.g., Poisson, and Laplace, which will be mainly used for testing and
further algorithm development, do not define a velocity field. I thought about
chaning the underlying design to separate the functionalitu provided via the
"base" class PDE into transport equations vs. non-transport equations, based on
whether a PDE defines a velocity field or not, respectively, but that would
require a lot of extra code and would complicate working with a vector of PDEs
singificantly. Instead, we keep the current design, i.e., there is only a single
PDE "base" class and the children have to now define a new member function
velocity(), which can optionally return an empty vector if a velocity field is
not defined by the PDE. As a result, we can keep the Carrier class general in
querying the velocity field, which, at this time, is only used to advance
particles across the mesh. To get the velocity field we query all PDEs
configured and use the last non-empty velocity vector as the velocity to advance
the particles with. As a result, by default, i.e., if the velocity() member
function for a newly defined PDE type is a do-nothing function, as, e.g., in
Poisson, i.e., returns an empty velocity vector, it is simply ignored. On the
other hand, a PDE type, e.g., Euler, CompNS, or AdvDiff, can define its own
arbitrary way of providing the velocity field. This keeps a lot of options open
for future research, e.g., adding a "velocity" field to a PDE that otherwise
makes no physical sense but which can be used to advance particles according to
some rule the PDE computes. This can be used, for example, to induce and
research load-imbalance (with particles) while computing only an elliptic
equation (that otherwise do not define a physical velocity field) without having
to solve a system of flow equations.