Correct boundary conditions collection
This commit fixes a couple of issues with the boundary conditions:

1. Previously the exodus reader was relying on "distribution factors" to find
out the number of nodes of a side set. This is okay (and it follows the example
in the exodus manual), but if there are no distribution factors in the mesh, it
fails to read the node IDs associated to a side set. A better way to read the
number of nodes that belong to a side set is via the
ex_get_side_set_node_list_len() call, which works even without distribution
factors.

2. Since the same node IDs can belong to the same side set repeatedly, there was
an ambiguity in how we applied boundary conditions to a node. This is now
changed to a different strategy: true values (indicating that a Dirichlet BC is
set on the given node for a given component) always overwrite false values
(indicating no BCs set on the given node for a given component).