Add regression tests for testing the asynchronous logic in inciter
This is follow-up to c36ceb4. These new tests only test whether the executable
was successfully finished a run and skip diffing any output. The tests excercise
the asynchronous logic in inciter's LinSysMerger (besides a lot of other
things). If the logic fails, hopefully the run deadlocks or an assert is
triggered in DEBUG mode. The tests sweep a number of CPUs and three different
virtualization settings.

Note that these tests do not excercise the full logic in LinSysMerger are the
test case do not have boundary conditions set, so not the full path of BC
verification is executed.