Overwrite diff instead of lump in LynSysMerger
Since the lumped mass matrix does not get recomputed if the mesh does not change
from a time step stage to another, it is wrong to overwrite it. So we now put
the low order solution into diff, the vector in which the mass diffusion
contribution to the right hand side was assembled. (That is recomputed in the
next time step stage as it depends on the previous-time solution.)