FCT now also correct in parallel and asynchronously
At f858caf, the FCT algorithm appeared correct, but was correct in serial and in
parallel only if the machine was not sufficiently disturbed, so asynchrony
kicked sufficiently in. As it turns out, several additions have been required to
make the asynchronous parallel algorithm correct.

As discussed in src/Inciter/carrier.ci, now there is no global synchronization
point necessary for the FCT algorithm. This is possible due to the way we
collect contributions to the data structures that store the antidiffusive
element contributions (AEC) applied to mesh nodes and the maxima and minima of
unknowns of all elements surrounding mesh nodes (ALW).  We now store the own and
those contributed by chares that operate on neighboring mesh chunks separate for
both AECs and ALWs. Besides AEC and ALW, computing and communicating the limited
antidiffusive element contributions (LIM) are done in an analogous fashion. This
enables overlapping computing and communicating these data and require no global
synchronization points.