Abstract particle communication from Carrier to Tracker
Particles in inciter are still rudimentary. The particles only carry position
and simply advanced by querying a velocity field from a PDE being integrated.
This commit abstracts away all particle functionality from Carrier to its own
class, Tracker, which is held by Carrier. While the functionality moved over is
largely unmodified, there is a couple of fixes to the particle communication
algorithm. Now it appears correct, tested with a simple rotation in a unit
cube, more testing is required.

Without having done any benchmarks, based on algorithmic complexity, the most
expensive part of the particle communication appears to be the removal of
particles. Since particle data is stored as tk::Data, which is really a
std::vector, removing an arbitrary element from the middle of the vector may
involve a lot of copying. This is mitigated by doing the copying only in a
single pass for removing a number of particles, but this is still rudimentary
and can be possibly improved on.

A potential improvement on the existing algorithm for particle communication
across chares is to temporarily leave holes in the particle array and add new
particles into existing holes instead of always adding new ones at the end of
the vector (as is currently done). This may also introduce a dependency between
different parts of the algorithm as they add and remove particles
asynchronously.

More details:

 * tk::Data::remove() has incorrectly modified the underlying data vector for
   multiple components. This has now been fixed and unit tests have been added.

 * tk::Data::add() has been renamed to push_back(), which is more consistent
   with the standard library lingo. Also the public interface to push_back() now
   delegates its work to a private function which is intentionally only
   overloaded on UnkEqComp data layout, because EqCompUnk would involve copying
   and could be much more expensive, and thus remains unimplemented on purpose.

 * As before, the number of particles per mesh elements is still hard coded into
   Carrier's constructor's initializer list, as well as the particle boundary
   conditions into Tracker::applyParBC(), which only work for a unit cube as is.

 * Carrier now also stores m_dt (the size of the time step), as Tracker::track()
   requires it to advance the particles, so m_dt must be available across
   Carrier's multiple member functions, e.g., apply(), rhs(), out().

 * It would be much more elegant to have Tracker store ParticleWriter. Though
   this could be done, Tracker would have to be templated on ParticleWriter's
   template argument. Additionally, Tracker still could only instantiate
   ParticleWriter if it somehow called back to Transporter, a single chare,
   (instead of an array or group). This is just not worth the code complexity.
   So ParticleWriter is left as is: created by Transporter, its proxy passed into
   Partitioner forwarded to Carrier and Tracker.

 * Note that Tracker still needs to invoke Charm++ proxy calls, for
   point-to-point communications as well as for broadcasts, so some of its member
   functions had to be templated, on both the holder object (Carrier) and the
   host of the holder (Transporter) so that the same code can be generated for
   classes different than Carrier and Transporter later.

 * The particle update previously has incorrectly used the difference between
   the previous and current time steps, instead of the velocity of the current
   time step. This has now been fixed, eliminating the need to store the
   previous velocity field in Carrier::m_up.

 * Another potential improvement to the particle communication is to not wait
   for those chares' anwsers, calling foundpar() and collectedpar(), which would
   not have anything to send, only for those that are sending something.
   Currently, both steps wait for answers regardless of the size/emptyness of
   the answer buffer.

 * No regression tests have been added as the particle boundary conditions are
   not general enough and the number of particles per cell has not been exposed
   to the user.