Only communicate nonempty BC data for LHS
This commit puts in a potentially temporary fix for a problem that
appears to only happen in extremely rare cases. The only regression test
failing was inciter:asynclogic_pe3 and only on a teamcity build machine.

For some reason, the all-to-all broadcast originating in
LinSysMerger:lhsbc() targeting LinSysMerger::bcval() would be sent
correctly from all 3 PEs, but only received on PEs 0 and 1. This yielded
a dead-lock lhsbc_complete() would never execute on PE 2. Note that this
only happens in the very first time step stage, as the LHS not
recomputed at this time. The particular regression test does not have
Dirichlet boundary conditions (BC), so the data to be sent and received
by these functions involved an empty std::unordered_map. I am not
certain why the not all receives would arrive, but for the case of no
Dirichlet BCs the solution put in by this commit appears to fix the
issue. The fix being put in is to skip the communication if the map is
empty, which is a correct and more efficient thing to do anyway.
However, if the BC map is not empty (for other problems), this problem
my resurface in the future -- hence this is a "potentially temporary
fix".