Serialize two reductions in inciter
We might still have an asynchronous logic bug. The symptom is highly
irreproducible. Only seems to yield a deadlock here and there on
variously loaded build agents during testing.

This commit serializes two reductions: (1) the one that signals the end
of computing the nodal volumes, originating from Carrier chares to
Transporter, and (2) the one that signals the end of sending off global
row indices, originating from LinSysMerger to Transporter. The hope is
that the deadlock goes away...

Also cleaned up the SDAG logic in Transporter.