Fix communication bug in partitioner
Only has happened rarely an on large problems, but part of the
communication was incorrectly leaving Partitioner::m_node incomplete
before it was used by Partitioner::reordered().

The problem was that a PE was already calling reordered() even though it
only sent its own portion of the global mesh node IDs to its
counterparts, without making sure it received all it had to. The fix is
to put in a global reduction (signal2host_distributed), signaling that a
PE has sent out and has made sure all of its contributions have been
received, followed by a broadcast to flatten() that now uses the
complete m_node (and not the potentially incomplete one).

Also put in some screen output to provide some feedback for larger
problems.