Update diagnostics computation
Since putting in FCT, LinSysMerger solves for the solution increment,
instead of the solution itself, an update was necessary in the way the
residuals were computed. This commit puts that in. In essence, we now
send the latest (limited) solution values to LinSysMerger, which then
does communication among other LinSysMerger branches on other PEs as
necessary, then computes the residuals the same way as they were before
(L1-norm).