Major revamp of Dirichlet BCs and minimum dt calculation
Details:

The way Dirichlet boundary conditions (BC) are set are revamped in a
major way.  When flux-corrected transport (FCT) was put in, we switched
to solving for the solution increment instead of the solution values.
This changes how Dirichlet BCs are applied. Previously the BC values had
to be enforced by putting in their prescribed nodal values in the rhs,
now the rhs only has to be zeroed at these locations.

The FCT algorithm required some minor modification for Dirichlet BCs.
The antidiffusive element contributions (AEC) now ignore the high order
solution increment at Dirichlet BC nodes. At the BC nodes this sets the
AEC equal to the negative mass diffusion used to add to the high order
right hand side to create the low order monotonic system. Since the high
order right hand side enforces the Dirichlet BCs, this yields the AEC
being equal to the low order increment times the inverse of the
assembled lumped mass matrix. (See also the detailed discussion in the
comments in FluxCorrector::aec(). Another change to the FCT algorithm is
how the AEC is applied at Dirichlet BC nodes. At BC nodes we always
apply the full AEC and ignore the limit coefficient. This has the
following advantages: (1) the Dirichlet BCs do not have to be set for
the low order system since the low order increment is completely ignored
at BC nodes (this also simplifies SDAG logic in LinSysMerger), and (2)
verification of the AEC remains the same and as useful as before, done
in FluxCorrector::verify(), which verifies (in serial) that AEC = dUh -
dUl.

The Dirichlet BCs are now set during setting the initial conditions for
all PDEs integrated. This makes sense as we set the ICs (and BCs) at
once during setup and the time stepping, computing only increments, only
modifies that solution.  Setting the ICs and BCs is done in
PDE::initialize() and can be outsourced to problem-specific functions.
This still needs some work as the ICs still need to be exposed to the
user.

The requirement to set the BCs at the same time as the the ICs are set
has led to another set of simplifications how Carrier chares prepare
their BC data structure, which also simplified the PDEs classes. The BC
map that associates a vector of pairs of booleans and boundary condition
values to mesh node IDs at which Dirichlet BCs are set, is now the only
data structure that Carrier chares as well as LinSysMerger group chares
hold. Though LinSysMerger does not require the BC values themselves,
only the node IDs and the booleans that determine whether a BC is set
for the given unknown or not, both store the same data structure for
simplicity. This is stored in Carrier::m_bc and constructed in
Carrier::bc(). Once m_bc is constructed, it is sent to LinSysMerger,
which broadcasts to all other PEs since setting Dirichlet BC on the LHS
matrix requires the knowledge of all BC nodes even the off-PE ones.
Carrier::m_bc is then not only used by LinSysMerger but also during
setting the initial conditions by all PDEs. The way this BC data
structure is constructed by Carrier is significantly simplified compared
to the previous solution, partly due to eliminating the BC verification,
discussed next.

Verification of the Dirichlet BCs, whether the (reordered, 'new') mesh
node IDs at which Dirichlet BCs are to be set match the (unordered,
'old') node IDs associated to (in the configuration file) the sideset in
the mesh file, are now taken out. This piece of code did only partial
verification of the BCs, i.e., it only verified the node IDs and not the
BC values, and it was just too complicated (involving complex
asynchronous logic and multiple reductions) for the limited
functionality it provided.

Now there are two different ways of specifying Dirichlet BCs:

1. As before:

   bc_dirichlet
     sideset <setid> <component> <value> end
   end

   This is useful for user-defined problems, e.g., setting BCs via
   sidesets matching the mesh file.

2. A new way:

   bc_dirichlet
     sideset <sideset ID list> end
   end

   This is useful for test problems, e.g., employing manufactured
   solutions, that specify BCs based on node coordinates using an algebraic
   problem-specific function. The reason the side set IDs need to be listed
   in the input file is to leverage the same machinery, i.e., reuse the
   same code, for matching the Dirichlet BCs in the mesh file and specified
   in the input file.

The sideset vector above, the list of numbers between the 'sideset' and
'end' keywords (for both styles), are now interpreted as a vector of
strings (for better or worse) during parsing. At this point this is a
list of integers and real numbers in the first case, and a list of
integers in the second case. These vectors of strings are processed in a
problem-specific way and converted to their correct types by the dirbc()
functions. For the first way of specifying the BC side set vector, we
now use std::stoi, std::stol, and std::atof, to convert the strings to
integer, long integer, and float, respectively, which should throw
exceptions if garbage is fed to them. This could probably use more
testing and error handling though.

The calculation of the minimum time step size has also been
significantly revamped. First, the skeleton for solving a Poisson
equation was taken out. The rest of the PDE types all fall into the
transport equations category, i.e., they can define a method (based on
physics and specific to problems) to compute the minimum time step size.
The calculation of the minimum dt is now delegated to PDEs as they are
in the best position to compute this. The length scale, necessary to
compute the time scale that defines the minimum dt, is now computed as
the cubic root of the cell volume which should perform better than the
minimum edge as was done previously. Also the compressible Euler
equation now properly takes the fluid velocity as well as the sound
speed into account when computing the dt based on advection. Depending
on the physics configured, if the Navier-Stokes equation is solved
instead of the Euler equations, i.e., if viscous forces and thermal
diffusion are present, the time scale is now computed taking these
properly into account.

Due to to the different way of computing the characteristic length scale
for computing the minimum dt (cubic root of cell volume instead of
minimum edge length) and the fix in Transporter::dt() that now correctly
takes the initial dt into account, some regression tests had to be
updated.

The slotted cylinder test case now starts differently rotated initial
conditions if multiple components are configured. We should/will add a
regression test for this, but now it is a good sign that the whole
asynchronous solution machinery using FCT holds together for a system of
equations.