Fix Dirichlet BCs on the LHS matrix in inciter
The way Dirichlet boundary conditions are applied to the left-hand-side
matrix in LinSysMerger was not finished correctly in b4e134b.

In short, setting Dirichlet BCs on the LHS was not correctly using the
new way b4e134b introduced storing its data structure. This commit fixes
that and also simplifies the logic by eliminating some communication (an
all-to-all on PEs yay!).

Also remove the progress sub-task indicators for assembly of the lhs
(from Transporter::m_progInit) and the rhs (from
Transporter::m_progStep).