Fix Dirichlet BC data structure communication
In LinSysMerger's setup (before time stepping can start) when
communicating the Dirichlet boundary conditions (BC) data structure so
that all PEs end up with the same data, previously each PE only
contributed if its m_bc data structure (supplied by those worker chares
that happen to be on their PE) was not empty. This was was done via an
all-to-all, but only those PEs contributed that had anything to send.
Also on a PE that had nothing to send the m_bc data was considered
complete at this point which is wrong, because it will still have to
receive the indices of those rows that other PEs own and set BCs on.
This is because even if a PE does not set a BC at a given row, it may
sitll have to zero its LHS matrix columns at the row IDs other PEs own
and set BCs on.

This commit removes the test for zero of the BC data structure, m_bc,
from LinSysMerger::charebc() and forces all PEs to contribute. (Not sure
why this was not an error, since only some of the PEs contributed to a
reduction.) Also, a PE that has nothing to send will now not signal
bc_complete(), only after it has received the result of the reduction
from all other PEs.

Thus this communication still remains to be an all-to-all, but now it is
supposed to be correct.