Fix svg scaling in doxygen output
For some reason doxygen stopped scaling svg images embedded in doxygen
pages and started displaying them page-wide. This commit fixes this by
moving the image scaling html commands to direct html instead of relying
on CSS. This is also nicer as it moves the image scaling configuration
to the place of inclusion.