Walker now samples from discrete data instead of relying on a fit
Previously (using a fit) and now (sampling from discrete data), Walker's
MixMassFractionBetaCoeffHydroTimeScaleHomDecay coefficient policy
achieves the same: provide a hydrodynamics time scale for the Monte
Carlo PDF simulation to help develop a binary mix model for
variable-density mixing in developing, i.e., transitional,
non-equilibrium, homogeneous Rayleigh-Taylor turbulence.

Using a fit was troublesome: if a good functional form for the direct
numerical simulation (DNS) data (at the given Atwood number and initial
configuration) could not be found, the time scale was hard to represent
and even if a good functional form could be found, a fit is still only a
fit.

With this commit we now have put in discrete functions of the inverse
hydrodynamics time scales as they evolve in DNS for various
configurations. For now, the simplest time scale (eps/k), defined for
equilibrium turbulence, is put in for Atwood numbers 0.05, 0.5, and 0.75
for initial conditions light << heavy (H), light = heavy (S), and light
>> heavy (L) for each Atwood number as extracted from DNS of homogeneous
Rayleigh-Taylor turbulence. All these 9 time scales together with the
steps for how the data is extracted from DNS and put in a suitable form
for source code are in src/DiffEq/HydroTimeScales.h.

The discrete functions are simply sampled at a given time as the
simulation proceeds using linear interpolation between the points. See
MixMassFracBetaCoeffHydroTimeScaleHomDecay::hydrotimescale().

Additional new time scales can be added and tested by adding new options
and time scales and new simulations can then be done running different
time scales simultaneously.

Selecting which time scale to use is exposed to the user via the keyword
'hydrotimescales', which can be used to associate one of the available
time scales to each of the components of the integrated system. (Run
'walker -H hydrotimescales' to see the help associated to this keyword
and the possible time scales that can be assigned.) The usual error
checking is done while parsing the hydrotimescales ...  end block: the
list must be the correct size, i.e., match the number of components
configured for the system, and the time scale labels must be selected
from the pool of available options for the hydrotimescales option
(tk::Toggle). Errors are emitted during parsing, as usual. See also
Control/Walker/Options/HydroTimeScales.h and Control/CommonGrammar.h
struct store_back_back_option.