Correct end-of-simulation logic in inciter
Previously, the logic to decide whether to continue time stepping in
inciter queried the time step stage as well as the number of iterations
and the max simulated time: if the time step stage was not the final in
a time step, the logic continued time stepping. It also continued if the
physical/simulated time was less than the max desired (term) or the
number of time step iterations was less then the max desired (nstep).

While the above is correct in theory, in becomes incorrect in the code
as this test, whether to continue or not, is done after the current time
step stage was already reset as part of getting ready for the
(potential) next stage. As a result, at the second time step stage,
continuation was always triggered, which started computing the right
hand sides again, etc. Though this was okay, but Transporter::finish()
could only be triggered after the first time step stage was finished,
which unnecessary.

This commit removes the condition that if the time step stage is the
first one, we continue. As a result, the end-of-simulation logic now is
only true if neither the iteration count nor the physical/simulation
time step has reached its desired maximum, independent of the time step
stage. As a result, the extra time step stage, before we can already
finish, does not happen, which is more of what was intended and was
would be expected by the user.

No regression tests change, since no output is done at half time step
stages.