Walker now also samples from discrete P/e data
In terms of functionality added, this commit is a follow-up to 337a2cc.

Walker's MixMassFractionBetaCoeffHydroTimeScaleHomDecay coefficient
policy now samples P/e from a number of built-in direct numerical
simulation (DNS) data for homogeneous Rayleigh-Taylor turbulence. P here
denotes the production of turbulent kinetic energy and e is the
dissipation rate of turbulent kinetic energy. Thus P/e from the
hydrodynamics field is available for Monte Carlo simulations, developing
a material mix model for variable-density mixing in developing, i.e.,
transitional, non-equilibrium, homogeneous Rayleigh-Taylor turbulence.

With this commit we now have put in discrete functions of P/e as they
evolve in DNS for various configurations. The configurations available
at this time are: Atwood numbers 0.05, 0.5, and 0.75 for initial
conditions light << heavy (H), light = heavy (S), and light >> heavy (L)
for each Atwood number as extracted from DNS of homogeneous
Rayleigh-Taylor turbulence. All these 9 P/e data together with the steps
for how the data is extracted from DNS and put in a suitable form for
source code are in src/DiffEq/HydroProductions.h.

The discrete functions are simply sampled at a given time as the
simulation proceeds using linear interpolation between the points. See
tk::sample() in Base/Table.[Ch].

Selecting which P/e data to use for a given component integrated is
exposed to the user via the keyword 'hydroproduction', which can be used
to associate one of the available data sets to each of the components of
the integrated system. (Run 'walker -H hydroproductions' to see the help
associated to this keyword and the possible P/e data sets that can be
assigned.) The usual error checking is done while parsing the
hydroproductions ...  end block: the list must be the correct size,
i.e., match the number of components configured for the system, and the
P/e labels must be selected from the pool of available options for the
hydroproductions option (tk::Toggle). Errors are emitted during parsing,
as usual. See also Control/Walker/Options/HydroProductions.h.