Fix error checking for time scales and P/e in walker
The error checking for the 'hydrotimescales' and 'hydroproductions'
vector blocks for HYDROTIMESCALE_HOMOGENEOUS_DECAY coefficients policy,
used in the MixMassFractionBeta DiffEq system was incorrect. It was not
an error if the user has not even specified one of these blocks, only if
the user has specified the block but with incorrect number of
components.

This change fixes this issue by checking for both extistence as well as
the correct size of these vectors.