Walker can now output extra statistics not from particles
This commit puts in the capability to walker to augment the textual
statistics output with extra columns containing data not originated from
ensemble particle statistics. This is useful for saving arbitrary
auxiliary non-particle data together with the particle-ensemble
statistics output, at the same time steps as the particle-statistics
use. Currently, we use this to output the inverse hydrodynamics time
scales and the production-to-dissipation ratios from direct numerical
simulation of homogeneous Rayleigh-Taylor turbulence that are available
as discrete tables embedded in the source. Since those discrete tables
are not necessarily sampled at the same temporal resolution as in the
Monte Carlo (MC) simulation might be desired by the user, plotting
statistics from MC scaled by data from the discrete tables so far
required to use the same time step sizes as the table data had to be
pasted from DNS which may use a different time step size. With this
commit using the same time step size for MC as in DNS (and the tables
from it) is no longer a requirement. The data output to the extra
columns are the same data used in the MC simulation sampled at the MC
time.

The labels for the extra columns come from the Keyword::name()
associated to the data table selected by the user.

Implementation details:

One way of adding these extra columns to the statistics output would
have been to simply make the sampled data as a new particle property and
drive it through the existing ensemble averaging procedure. The problem
with that approach is that the sampled data is a single real value (per
table per component) and the same for all particles in the ensemble.
This is a waste of memory and computation.

Another way of doing it is to move the data sampled at its MC usage site
(i.e., in the coefficients policy) all the way back to Distributor
(single object on a single PE). Since sampling the tables is done deep
down in a particular coefficient policy for a differential equation
object, it seemed wasteful to try to send that data back to where the
data needs to be output to file, which is at the highest point in
Distributor::outStat(). This would have required a lot of interface
changes as well as additional communication including collectives, which
is a waste as well.

Instead, Distributor prepares the tables in its constructor, more
precisely in Distributor::info(). This preparation involves
instantiating all discrete tables in a vector (done on a single PE only)
and then sampling them all immediately before the data needs to be
written out to file in outStat(). While the table preparation /
instantiation step is necessarily DiffEq-specific, sampling the tables
is not. The instantiation is done in DiffEqStack::tables() which calls
DiffEqStack::createTable(). This latter is responsible for finding out
which differential equation system type uses tables and which ones. At
this time, only MixMassFractionBeta's HYDROTIMESCALE_HOMOGENEOUS_DECAY
does, but others can be added in the future.

This solution separates the DiffEq-specific table-instantiation and the
generic sampling steps and most importantly completely avoids extra
communication. The price for this close-to-zero cost is some minor code
and data duplication.

Since the result of the sampling of the tables is required at two
different points (at computation inside the coefficients policy and at
the statistics output), a couple of lines of code and the data for the
tables are duplicated. The duplicate code in the instantiation of the
tables is done in the DiffEq (MixMassFractionBeta) constructor as well
as DiffEq::createTables. Storage of the tables is also duplicated: in
the entry in g_diffeqs as well as in Distributor::m_tables. Sampling is
also duplicated: once in the coefficients policy and once in
Distributor::outStat(). This also avoids changing the API for all other
differential equation types and policies. Neither of the code nor the
data duplication is really a concern. The code is simply a linear
interpolation between two points of the discrete tables - this is
immeasurably smaller than anything else being done for Monte Carlo. The
data in the tables is also extremely small compared to the memory usage
of the ensemble particles.