Several fixes
* Add forgotten const qualifiers to Input& for some of the newly
converted actions in CommonGrammar.h.

* Pull over escape/escaper from PEGTL v0.32 to extract std::string from
pegtl::string inside struct keyword.

* Parser location disabled for now.

* Convert meshconv's grammar.

This now compiles, but the abstraction of rules and actions are not
correct yet, so at runtime thought it triggers the correct actions for,
e.g., a command line argument '-i', the matched input is empty for some
reason.

Work in progres...