Fix PEGTL (as own-TPL) install path
PEGTL as our own TPL (via git subtree) now should point to

git@github.com:ColinH/PEGTL.git

instead of

git@github.com:quinoacomputing/PEGTL-0.32.git

With 15186011ee0 our git-subtree PEGTL points to Colin's master branch
as we will start using the latest features (put in as a response with
our conversion to the new PEGTL).

We continue searching for a system-wide PEGTL and if that is found, we
use it.

As always forcing our own PEGTL (instead of the system's) can be done
via the -DNO_SYSTEM_PEGTL=true cmake argument of the TPL build.