Make RNGSSE2 optional
Building RNGSSE2 is now conditional on the detected system architecture.
If the arch is not x86, RNGSSE2 cannot build, so we don't build it and
conditionally remove it from the code, including its unit and regression
tests.

In the TPL cmake run, building of RNGSSE2 can be overridden (even on x86
arch) by the -DNO_RNGSSE2=true cmake arg. This will be useful for
testing the absence of RNGSSE2 on x86 architectures, e.g., during CI.

If none of the three RNG libs (MKL, RNGSSE2, and Random123) are found
when attempting to build quinoa, we error out during the cmake run.

Couple of regression tests needed minor adjustments: one slightly
loosening the relative tolerance, another one to switch the RNG used.