Rely on Trilinos-installed find_package for Zoltan2, Exodus, and Exodiff
Notes:

 * The TPL cmake-run will attempt to find the Trilinos libraries
quietly, which means that if it fails, we build our own Trilinos. Cmake
searches at system paths as well as at those given by <package>_DIR
cmake or environment variables. Note that during the TPL cmake run these
find_package calls are not only quiet (i.e., failure is okay) but also
do not require the packages to be found, i.e., optional (by default).

* The code cmake-run requires the find_package calls to succeed. Not
only do they have to succeed because they find their respected
libraries, but also because their <package>Config.cmake or
Find<package>cmake are available. CMAKE_PREFIX_PATH is populated by
TPL_DIR which is where we install our TPLs if they were not found during
the TPL cmake run. By this time, the <package>Config.cmake files are
required to be available. If we installed Trilinos ourselves, they are
at their respective directories at ${TPL_DIR}/lib/cmake/<package>. This
is found by cmake since we set CMAKE_PREFIX_PATH and cmake's internal
logic finds them relative to TPL_DIR. If we did not install Trilinos
ourselves because the TPL cmake-run found it on the system, it will
(should?) find the <package>Config.cmake again during the code cmake-run
too. If we did not install Trilinos ourselves because the TPL cmake-run
found it at a path given by either a cmake or an environment variable
<package>_DIR, these must be available to the code cmake-run too,
otherwise an error is generated. Therefore this setup assumes that
either (1) the same set of -D<package>_DIR=... cmake variables are
passed to both the TPL and the code cmake-runs or (2) both the TPL and
the cmake-runs happen in the same environemnt in which the same
environment variables are set. This latter can be reasonably assumed.