Remove hash-set of node IDs in partitioner
The hash set (std::unordered_set in Partitioner::m_sid) being removed
used to store the global mesh node IDs during parallel distributed mesh
node reordering. The reason for its existence is a search required in
Partitioner::query(), which is more efficient using a hash set. However,
it uses additional memory, which blows out of 128GB/node on a job
running with a 794M-cell mesh on 600 nodes (20K CPU cores). Since this
data is really duplicate, this commit removes it and replaces the search
with a linear search over another data array (m_id) that is a
std::vector. This enables running but the search is potentially very
inefficient.

Another solution might be to replace m_id with a hash-set (instead of a
vector), but that data must also be communicated (and has-sets may be
slower to communicate than vectors) so there is a trade-off,
nevertheless should be tried, because that may similarly reduce memory
usage.