A better solution for the problem 0a0b357d3c2 solves
This commit switches Partitioner::m_id from std::vector to a(n ordered)
std::set. As a result, Partitioner::mask() does not have to do a linear
search for the node IDs, but can do one with an n*log(n) cost. This
still saves memory because we still get rid of the original data
duplication 0a0b357d3c2 discusses, but the search is faster.

Run a couple of tests to quantify the improvement: the linear search
costs a total runtime of 37:12 (min:sec) of a 794M-cells mesh on 14K
CPUs, while the n*log(n) search yields 13:57. This is the effect on the
total(!) runtime of a single loop during pre-rpocessing (done once)
during setup (parallel distributed mesh node reordering). Another
similar run, (same problem, but different number of CPUs), is 7.2K CPUs
the difference in total runtime is 53:07 vs. 13:02, that's 4x.

The time-stepping is not affected by this.