Simplify Partitioner's query() and mask()
This commit removes generating and communicating a vector of integers
during mesh node reordering. The information contained in the removed
vector (own and rec on the sending and receiving sides, respectively) is
now obtained from the std::set arriving in ch and queried by testing if
the set is empty. This should also help on memory usage.